+++ src/lwc/geConstants/geConstants.js
@@ -9,9 +9,21 @@ const HTTP_CODES = Object.freeze({
     Request_Timeout: 408,
 });
 
+const PAYMENT_TRANSACTION_STATUS_ENUM = Object.freeze({
+    PENDING: 'PENDING',
+    AUTHORIZED: 'AUTHORIZED',
+    CANCELED: 'CANCELED',
+    CAPTURED: 'CAPTURED',
+    DECLINED: 'DECLINED',
+    NONRETRYABLEERROR: 'NONRETRYABLEERROR',
+    RETRYABLEERROR: 'RETRYABLEERROR',
+    REFUNDISSUED: 'REFUNDISSUED'
+});
+
 export {
     DISABLE_TOKENIZE_WIDGET_EVENT_NAME,
     LABEL_NEW_LINE,
     WIDGET_TYPE_DI_FIELD_VALUE,
     HTTP_CODES,
+    PAYMENT_TRANSACTION_STATUS_ENUM
 };
